\name{BNPqte-package}
\alias{BNPqte-package}
\alias{BNPqte}
\docType{package}
\title{
\packageTitle{BNPqte}
}
\description{
\packageDescription{BNPqte}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BNPqte}
\packageIndices{BNPqte}
~~ An overview of how to use the ~~
~~ package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{BNPqte}

Maintainer: \packageMaintainer{BNPqte}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
